+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.4 - UNRELEASED
 
+It is now possible to set a prefix for a job; typically used to identify the job (much like a thread name).
+
 ## 0.1.3 - 05 Feb 2016
 
 Added `*terminal-type*` Var, to override default terminal type passed to the `tput` command.